books/serializers.py
from rest_framework import serializers
from .models import Publisher

class PublisherSerializer(serializers.ModelSerializer):
    """Serializer to map the Model instance into JSON format."""

    class Meta:
        """Meta class to map serializer's fields with the model fields."""
        model = Publisher
        fields = ('pk', 'name', 'address', 'city', 'state_province', 'country', 'website')