requirements.txt
Django==1.11.14
djangorestframework==3.8.2
pytz==2018.5
wsgiref==0.1.2
requests==2.19.1
